Lintcheck: Refactor structs and only take one version per crate
